Smarter image shrinking
Limit maximum image size almost fill the entire viewport so it doesn't
ever get too overwhelming

This needed an additional inner wrapper for the figure, so that we could
set inline-block (to get the border to wrap around the image) +
text-align: center.

If it wasn't inline-block, the border would stretch across the whole
page even after we shrink the image.